Fix broker queues API with Redis.
With non-AMPQ brokers `get_manager_client` returns `None` for the client
which obviously doesn't have the `is_alive()` method. So we need to
check if the connection driver type is AMQP otherwise
`/v1/broker/queues` returns 503 with a `broker_not_reachable` error. Fix
the same issue with `purge_queue()` which I haven't encountered.